Moved button drawing and layout into button.dart and own widget wrapper
