+++ lib/session.dart
@@ -19,8 +19,6 @@
 /// Main session view
 
 import 'dart:async';
-import 'dart:math' show min, max;
-import 'dart:ui' as ui;
 
 import 'package:embla/loc.dart';
 import 'package:flutter/cupertino.dart';
@@ -44,6 +42,7 @@ import './menu.dart' show MenuRoute;
 import './prefs.dart' show Prefs;
 import './jsexec.dart' show JSExecutor;
 import './theme.dart';
+import './button.dart';
 import './util.dart' show readServerAPIKey;
 import './version.dart' show getClientType, getVersion, getUniqueIdentifier;
 
@@ -55,47 +54,15 @@ const kNoInternetMessage = 'Ekki næst samband við netið.';
 const kNoMicPermissionMessage =
     'Ekki tókst að hefja talgreiningu. Emblu vantar heimild til að nota hljóðnema.';
 
-// Waveform configuration
-const int kWaveformNumBars = 12; // Number of waveform bars drawn
-const double kWaveformBarMarginRatio = 0.22; // Spacing between waveform bars as proportion of width
-const double kWaveformDefaultSampleLevel = 0.05; // Slightly above 0 looks better
-const double kWaveformMinSampleLevel = 0.025; // Hard limit on lowest level
-const double kWaveformMaxSampleLevel = 0.95; // Hard limit on highest level
-
 // Animation framerate
 const int msecPerFrame = (1000 ~/ 24);
 const Duration durationPerFrame = Duration(milliseconds: msecPerFrame);
-
-// Logo animation status
-const kFullLogoFrame = 99;
-int currFrame = kFullLogoFrame;
-
-// Session button size (proportional to width/height)
-const kRestingButtonPropSize = 0.58;
-
-// Session button accessibility labels
-const kRestingButtonLabel = 'Tala við Emblu';
-const kExpandedButtonLabel = 'Hætta að tala við Emblu';
 
 // Hotword detection accesibility labels
 const kDisableHotwordDetectionLabel = 'Slökkva á raddvirkjun';
 const kEnableHotwordDetectionLabel = 'Kveikja á raddvirkjun';
 
 BuildContext? sessionContext;
-
-// Samples (0.0-1.0) used for waveform animation
-List<double> audioSamples = populateSamples();
-
-List<double> populateSamples() {
-  return List.filled(kWaveformNumBars, kWaveformDefaultSampleLevel, growable: true);
-}
-
-void addSample(double level) {
-  while (audioSamples.length >= kWaveformNumBars) {
-    audioSamples.removeAt(0);
-  }
-  audioSamples.add(level < kWaveformDefaultSampleLevel ? kWaveformDefaultSampleLevel : level);
-}
 
 // Main widget for session view
 class SessionRoute extends StatefulWidget {
@@ -273,7 +240,7 @@ class SessionRouteState extends State<SessionRoute> with TickerProviderStateMixi
       setState(() {
         text = '';
         imageURL = null;
-        audioSamples = populateSamples();
+        Waveform().setDefaultSamples();
         animationTimer?.cancel();
         animationTimer = Timer.periodic(durationPerFrame, (Timer t) => ticker());
       });
@@ -376,14 +343,9 @@ class SessionRouteState extends State<SessionRoute> with TickerProviderStateMixi
 
   @override
   Widget build(BuildContext context) {
-    // Session button properties depending on whether session is active
     sessionContext = context;
-    final bool active = session.isActive();
-    const double prop = kRestingButtonPropSize;
-    final double buttonSize = MediaQuery.of(context).size.width * prop;
-    final String buttonLabel = active ? kRestingButtonLabel : kExpandedButtonLabel;
 
-    // Hotword toggle button properties depending on whether hw detection is enabled
+    // Hotword toggle button properties depend on whether hotword detection is enabled
     final bool hwActive = Prefs().boolForKey('hotword_activation');
     final String hotwordIcon = hwActive ? 'mic' : 'mic-slash';
     final String hotwordLabel =
@@ -449,38 +411,19 @@ class SessionRouteState extends State<SessionRoute> with TickerProviderStateMixi
               )));
     }
 
-    // Generate widget tree for the session button below the text area
-    Widget sessionButtonWidget() {
-      return Padding(
-          padding: const EdgeInsets.only(bottom: 30, top: 30),
-          child: Center(
-              child: Semantics(
-                  label: buttonLabel,
-                  child: GestureDetector(
-                      onTap: toggle,
-                      child: SizedBox(
-                        width: buttonSize,
-                        height: buttonSize,
-                        // Session button uses custom painter to draw the button
-                        child: CustomPaint(painter: SessionButtonPainter(context, session))
-                            .animate(target: session.isActive() ? 1 : 0)
-                            .scaleXY(end: 1.20, duration: 100.ms),
-                      )))));
-    }
-
     return Scaffold(
-      // Top nav bar
+      // Top navigation bar
       appBar: AppBar(
           bottomOpacity: 0.0,
           elevation: 0.0,
-          // Toggle hotword activation button
+          // Toggle hotword activation button (left)
           leading: Semantics(
               label: hotwordLabel,
               child: IconButton(
                 icon: ImageIcon(img4theme(hotwordIcon, context)),
                 onPressed: toggleHotwordActivation,
               )),
-          // Hamburger menu button
+          // Hamburger menu button (right)
           actions: <Widget>[
             Semantics(
                 label: 'Sýna valblað',
@@ -493,144 +436,9 @@ class SessionRouteState extends State<SessionRoute> with TickerProviderStateMixi
           // Session text widget
           Expanded(flex: 6, child: scrollableTextAreaWidget()),
           // Session button widget
-          Expanded(flex: 8, child: sessionButtonWidget()),
+          Expanded(flex: 8, child: SessionButtonWidget(context, session, toggle)),
         ],
       ),
     );
-  }
-}
-
-// This is the drawing code for the session button
-class SessionButtonPainter extends CustomPainter {
-  late final EmblaSession session;
-  late final BuildContext context;
-  SessionButtonPainter(this.context, this.session);
-
-  // Draw the three circles that make up the button
-  void drawCircles(Canvas canvas, Size size) {
-    final radius = min(size.width, size.height) / 2;
-    final center = Offset(size.width / 2, size.height / 2);
-
-    final List<Color> circleColors = circleColors4Context(sessionContext);
-
-    // First, outermost circle
-    var paint = Paint()..color = circleColors[0];
-    canvas.drawCircle(center, radius, paint);
-
-    // Second, middle circle
-    paint = Paint()..color = circleColors[1];
-    canvas.drawCircle(center, radius / 1.25, paint);
-
-    // Third, innermost circle
-    paint = Paint()..color = circleColors[2];
-    canvas.drawCircle(center, radius / 1.75, paint);
-  }
-
-  // Draw current logo animation frame
-  void drawLogoFrame(Canvas canvas, Size size, int fnum) {
-    if (animationFrames.isEmpty) {
-      dlog('Animation frame drawing failed. No frames loaded.');
-      return;
-    }
-    final ui.Image img = animationFrames[fnum];
-    // Source image rect
-    final Rect srcRect = Rect.fromLTWH(0, 0, img.width.toDouble(), img.height.toDouble());
-
-    // Destination rect centered in canvas
-    final double sw = size.width.toDouble();
-    final double sh = size.height.toDouble();
-    const double prop = 2.4;
-    final double w = sw / prop;
-    final double h = sh / prop;
-    final Rect dstRect = Rect.fromLTWH(
-        (sw / 2) - (w / 2), // x
-        (sh / 2) - (h / 2), // y
-        w, // width
-        h); // height
-    canvas.drawImageRect(img, srcRect, dstRect, Paint());
-  }
-
-  // Draw audio waveform
-  void drawWaveform(Canvas canvas, Size size) {
-    // Generate square frame to contain waveform
-    final double w = size.width / 2.0;
-    final double xOffset = (size.width - w) / 2;
-    final double yOffset = (size.height - w) / 2;
-    final Rect frame = Rect.fromLTWH(xOffset, yOffset, w, w);
-
-    final double margin = (size.width * kWaveformBarMarginRatio) / (kWaveformNumBars - 1);
-    final double totalMarginWidth = (kWaveformNumBars * margin) - margin;
-
-    final double barWidth = (frame.width - totalMarginWidth) / kWaveformNumBars;
-    final double barHeight = frame.height / 2;
-    final double centerY = (frame.height / 2);
-
-    // Colors for the top and bottom waveform bars
-    final topPaint = Paint()..color = topWaveformColor;
-    final bottomPaint = Paint()..color = bottomWaveformColor;
-
-    // Draw audio waveform bars based on audio sample levels
-    for (int i = 0; i < audioSamples.length; i++) {
-      // Clamp signal level
-      final double level =
-          min(max(kWaveformMinSampleLevel, audioSamples[i]), kWaveformMaxSampleLevel);
-
-      // Draw top bar
-      final Rect topRect = Rect.fromLTWH(
-          i * (barWidth + margin) + (margin / 2) + xOffset, // x
-          barHeight - (level * barHeight) + yOffset, // y
-          barWidth, // width
-          level * barHeight); // height
-      canvas.drawRect(topRect, topPaint);
-
-      // Draw circle at end of bar
-      canvas.drawCircle(
-          Offset(i * (barWidth + margin) + barWidth / 2 + (margin / 2) + xOffset,
-              barHeight - (level * barHeight) + yOffset), // offset
-          barWidth / 2, // radius
-          topPaint);
-
-      // Draw bottom bar
-      final Rect bottomRect = Rect.fromLTWH(
-          i * (barWidth + margin) + (margin / 2) + xOffset, // x
-          centerY + yOffset, // y
-          barWidth, // width
-          level * barHeight); // height
-      canvas.drawRect(bottomRect, bottomPaint);
-
-      // Draw circle at end of bottom bar
-      canvas.drawCircle(
-          Offset(i * (barWidth + margin) + barWidth / 2 + (margin / 2) + xOffset,
-              centerY + (level * barHeight) + yOffset), // offset
-          barWidth / 2, // radius
-          bottomPaint);
-    }
-  }
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    // We always draw the circles
-    drawCircles(canvas, size);
-
-    // Draw waveform bars during microphone input
-    if (session.state ==
-            EmblaSessionState.listening /*||
-        session.state == EmblaSessionState.starting*/
-        ) {
-      drawWaveform(canvas, size);
-    }
-    // Draw logo animation during answering phase
-    else if (session.state == EmblaSessionState.answering) {
-      drawLogoFrame(canvas, size, currFrame);
-    }
-    // Otherwise, draw non-animated Embla logo
-    else {
-      drawLogoFrame(canvas, size, kFullLogoFrame); // Always same frame
-    }
-  }
-
-  @override
-  bool shouldRepaint(SessionButtonPainter oldDelegate) {
-    return true;
   }
 }
