+++ lib/smart/add_connection.dart
@@ -15,7 +15,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-// mDNS scan route
+/// Add connection route
 
 import 'dart:core';
