+++ lib/smart/mdns_scan.dart
@@ -15,7 +15,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-// mDNS scan route
+/// mDNS scan route
 
 import 'package:flutter/material.dart';
